Edit builder

+++ grails-app/services/com/everlution/StepService.groovy
@@ -15,15 +15,17 @@ abstract class StepService implements IStepService {
     @Transactional
     void delete(Serializable id) {
         def step = get(id)
-        def links = linkService.getLinks(step.id, step.project)
-        links.each {
-            linkService.delete(it.id)
+        if (step) {
+            def links = linkService.getLinks(step.id, step.project)
+            links.each {
+                linkService.delete(it.id)
+            }
+            step.delete()
         }
-        step.delete()
     }
 
     /**
-     * finds all steps in a project
+     * finds all builder steps in a project
      */
     @Transactional
     SearchResult findAllByProject(Project project, Map args) {
@@ -33,7 +35,7 @@ abstract class StepService implements IStepService {
     }
 
     /**
-     * finds all steps in the project with a name
+     * finds all builder steps in the project with a name
      * that contains the string
      * @param name - the string to search
      */
